#17 Fix the TF-reassignment requantization to output frames.
The problem was caused by totally wrong way for requantization! There are input frames in the spectrogram which may be overlapping. From each cell we compute the corrected position in time and frequency. Then we quantize this set of points into another grid, this time with non-overlapping time frames using np.histogram2d(). This was provided however by parameters related to the input overlapping frames.

The solution was to properly specify the output frames and and provide correct parameters for np.histogram2d().

After the fix TF-reassignment works like charm. Plain frequency reassignment has to be implemented the same way (not on the input frames), just omitting the time corrections. By providing uncorrected frequencies we can also do time reassignment. Finally we can perform requantization without any TF correction.

The only remaining problem is if we don't use time correction and the input frames are larger than the output frames the energy is not distributed evenly and there are gaps in time. This is due to the fact that in both cases the value represents something else - point or area. The proper way would be to resample the energy into the output frames according to the overlap and possibly weighted by the FFT window.